fix(bitmaps): remove some awkward and low-contrast icons
This will force the go- navigation arrows and the zoom controls to
fall back to using symbolic icons, which are dynamically recolored
by the theme. This will improve contrast in applications which use
the icons. It also obviates the need for adding a contrast-stroke
to these icons, which is awkward and doesn't look as good.